requirements.txt
beautifulsoup4==4.13.4
cssutils==2.11.1
more-itertools==10.7.0
pillow==11.3.0
soupsieve==2.7
typing_extensions==4.14.1
